refactor(contentful): split StringFieldImporter into reducer & updater
