+++ packages/botonic-plugin-contentful/tests/tools/l10n/csv-import.test.ts
@@ -1,15 +1,14 @@
 import path from 'path'
+import { CsvImport, Record } from '../../../src/tools/l10n/csv-import'
 import { anything, instance, mock, when } from 'ts-mockito'
-
 import * as cms from '../../../src/cms'
 import { ContentFieldType } from '../../../src/manage-cms/fields'
-import { CsvImport, Record } from '../../../src/tools/l10n/csv-import'
-import { StringFieldImporter } from '../../../src/tools/l10n/string-field-importer'
+import { ImportRecordReducer } from '../../../src/tools/l10n/import-updater'
 
 const FIXTURES_BASE = path.resolve(__dirname, '__fixtures__')
 
 test('TEST: CsvImport read text, URL & carousel', async () => {
-  const mockFieldImporter = mock<StringFieldImporter>()
+  const mockFieldImporter = mock<ImportRecordReducer>()
   const readLines: Record[] = []
   let flushCount = 0
   when(mockFieldImporter.consume(anything())).thenCall((record: Record) => {
